+++ packages/uniforms-bridge-simple-schema-2/src/SimpleSchema2Bridge.ts
@@ -1,6 +1,7 @@
 import SimpleSchema from 'simpl-schema';
 import cloneDeep from 'lodash/cloneDeep';
 import invariant from 'invariant';
+import memoize from 'lodash/memoize';
 import { Bridge, joinName } from 'uniforms';
 
 export default class SimpleSchema2Bridge extends Bridge {
@@ -10,6 +11,11 @@ export default class SimpleSchema2Bridge extends Bridge {
     super();
 
     this.schema = schema;
+
+    // Memoize for performance and referential equality.
+    this.getField = memoize(this.getField);
+    this.getSubfields = memoize(this.getSubfields);
+    this.getType = memoize(this.getType);
   }
 
   getError(name, error) {
@@ -166,12 +172,14 @@ export default class SimpleSchema2Bridge extends Bridge {
 
   getValidator(options = { clean: true, mutate: true }) {
     const validator = this.schema.validator(options);
-
-    // Clean mutate its argument, even if mutate is false.
-    if (options.clean) {
-      return model => validator(cloneDeep({ ...model }));
-    }
-
-    return validator;
+    return model => {
+      try {
+        // Clean mutate its argument, even if mutate is false.
+        validator(options.clean ? cloneDeep({ ...model }) : model);
+        return null;
+      } catch (error) {
+        return error;
+      }
+    };
   }
 }