+++ packages/uniforms/src/BaseForm.tsx
@@ -1,9 +1,9 @@
 import React, { Component, SyntheticEvent } from 'react';
-import cloneDeep from 'lodash/cloneDeep';
+import clone from 'lodash/clone';
 import get from 'lodash/get';
 import isFunction from 'lodash/isFunction';
 import omit from 'lodash/omit';
-import set from 'lodash/set';
+import setWith from 'lodash/setWith';
 
 import { Bridge } from './Bridge';
 import { ChangedMap, Context, DeepPartial, ModelTransformMode } from './types';
@@ -15,7 +15,7 @@ export type BaseFormProps<Model> = {
   autosave: boolean;
   autosaveDelay: number;
   disabled?: boolean;
-  error?: any;
+  error: any;
   id?: string;
   label: boolean;
   model: DeepPartial<Model>;
@@ -25,9 +25,7 @@ export type BaseFormProps<Model> = {
   ) => DeepPartial<Model>;
   noValidate: boolean;
   onChange?(key: string, value: any): void;
-  onSubmit?(model: DeepPartial<Model>): any;
-  onSubmitFailure?(result: any): void;
-  onSubmitSuccess?(result: any): void;
+  onSubmit(model: DeepPartial<Model>): any;
   placeholder?: boolean;
   schema: Bridge;
   showInlineError?: boolean;
@@ -49,9 +47,11 @@ export class BaseForm<
   static defaultProps = {
     autosave: false,
     autosaveDelay: 0,
+    error: null,
     label: true,
     model: Object.create(null),
     noValidate: true,
+    onSubmit() {},
   };
 
   constructor(props: Props) {
@@ -184,13 +184,18 @@ export class BaseForm<
     if (this.mounted) {
       const keys = changedKeys(key, value, get(this.getModel(), key));
       if (keys.length !== 0) {
-        this.setState(state => ({
-          changed: true,
-          changedMap: keys.reduce(
-            (changedMap, key) => set(changedMap, key, {}),
-            cloneDeep(state.changedMap),
-          ),
-        }));
+        this.setState(state =>
+          // If all are already marked, we can skip the update completely.
+          state.changed && keys.every(key => !!get(state.changedMap, key))
+            ? null
+            : {
+                changed: true,
+                changedMap: keys.reduce(
+                  (changedMap, key) => setWith(changedMap, key, {}, clone),
+                  clone(state.changedMap),
+                ),
+              },
+        );
       }
     }
 
@@ -234,24 +239,15 @@ export class BaseForm<
       event.stopPropagation();
     }
 
-    const result =
-      this.props.onSubmit && this.props.onSubmit(this.getModel('submit'));
-
-    // Set the `submitting` state only if onSubmit is async so we don't cause an unnecessary re-render
-    let submitting: Promise<any>;
-    if (isPromiseLike(result)) {
-      this.setState({ submitting: true });
-      submitting = result.finally(() => {
-        this.setState({ submitting: false });
-      });
-    } else {
-      submitting = Promise.resolve(result);
+    const result = this.props.onSubmit(this.getModel('submit'));
+    if (!(result instanceof Promise)) {
+      return Promise.resolve(result);
     }
 
-    return submitting.then(
-      this.props.onSubmitSuccess,
-      this.props.onSubmitFailure,
-    );
+    this.setState({ submitting: true });
+    return result.finally(() => {
+      this.setState({ submitting: false });
+    });
   }
 
   render() {
@@ -261,8 +257,4 @@ export class BaseForm<
       </context.Provider>
     );
   }
-}
-
-function isPromiseLike(value: any): value is Promise<any> {
-  return !!value && isFunction(value.then);
 }